Move login command into own file
